Fix credit behaviour for mobile, update structure, add key
